+++ src/components/main-app/__tests__/component.test.js
@@ -10,6 +10,7 @@ import MainApp from '../component'
 import EstimationCard from '../../estimation-card'
 import BoardsList from '../../boards-list'
 import MembersList from '../../members-list'
+import Notification from '../../notification'
 
 describe('Component/MainApp', () => {
   const props = {
@@ -40,10 +41,31 @@ describe('Component/MainApp', () => {
 
   test('should render without throwing an error', () => {
     const wrapper = shallow(<MainApp {...props} />).dive()
-    expect(wrapper.find(EstimationCard).length).toEqual(1)
+    expect(wrapper.find(EstimationCard).length).toEqual(0) // only visible after toggle
+    expect(wrapper.find(Notification).length).toEqual(0) // only visible when property appErrors is valid
     expect(wrapper.find(BoardsList).length).toEqual(1)
     expect(wrapper.find(MembersList).length).toEqual(1)
     expect(wrapper).toMatchSnapshot()
+  })
+
+  test('should show EstimationCard when toggled', () => {
+    const wrapper = shallow(<MainApp {...props} />).dive()
+    expect(wrapper.find(EstimationCard).length).toEqual(0)
+    wrapper.setState({
+      showEstimationCard: true,
+    })
+    wrapper.update()
+    expect(wrapper.find(EstimationCard).length).toEqual(1)
+  })
+
+  test('should show Notification when appErrors is a valid array', () => {
+    const wrapper = shallow(<MainApp {...props} />).dive()
+    expect(wrapper.find(Notification).length).toEqual(0)
+    wrapper.setProps({
+      appErrors: ['example error'],
+    })
+    wrapper.update()
+    expect(wrapper.find(Notification).length).toEqual(1)
   })
 
   test('should render an LinearProgress when isLoading is true', () => {