+++ src/components/notification/index.js
@@ -1,0 +1,70 @@
+// Docs https://material-ui.com/demos/snackbars/
+import React from 'react'
+import PropTypes from 'prop-types'
+import { withStyles } from '@material-ui/core/styles'
+import Snackbar from '@material-ui/core/Snackbar'
+import IconButton from '@material-ui/core/IconButton'
+import CloseIcon from '@material-ui/icons/Close'
+
+const styles = theme => ({
+  close: {
+    width: theme.spacing.unit * 4,
+    height: theme.spacing.unit * 4,
+  },
+})
+
+class Notification extends React.Component {
+  constructor(props) {
+    super(props)
+    this.state = {
+      open: typeof props.message !== 'undefined' || false,
+    }
+  }
+
+  handleClose = (event, reason) => {
+    if (reason === 'clickaway') {
+      return
+    }
+
+    this.setState({ open: false })
+  }
+
+  render() {
+    const { classes, message } = this.props
+    return (
+      <div>
+        <Snackbar
+          anchorOrigin={{
+            vertical: 'bottom',
+            horizontal: 'right',
+          }}
+          open={this.state.open}
+          autoHideDuration={10000}
+          onClose={this.handleClose}
+          ContentProps={{
+            'aria-describedby': 'message-id',
+          }}
+          message={<span id="message-id">{message}</span>}
+          action={[
+            <IconButton
+              key="close"
+              aria-label="Close"
+              color="inherit"
+              className={classes.close}
+              onClick={this.handleClose}
+            >
+              <CloseIcon />
+            </IconButton>,
+          ]}
+        />
+      </div>
+    )
+  }
+}
+
+Notification.propTypes = {
+  classes: PropTypes.object.isRequired, // eslint-disable-line
+  message: PropTypes.string.isRequired,
+}
+
+export default withStyles(styles)(Notification)